Use the SDL_GameController API for axes values
The GameController API fixes the axis mappings on most controllers, but
screws the button and POV mappings.

To fix this issue, we only use the GameController API for registering
axis movement, while using the classic joystick API for buttons and
POVs.

In other words, we get the best from both worlds.